+++ src/test/java/backend/likelion/todos/member/MemberControllerTest.java
@@ -5,7 +5,6 @@ import static org.springframework.http.HttpStatus.CONFLICT;
 import static org.springframework.http.HttpStatus.CREATED;
 
 import backend.likelion.todos.ApiTest;
-import backend.likelion.todos.member.request.SignupRequest;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
